fix(ai): remove integer type from AI schema (Anthropic rejects min/max on integers)
Anthropic structured output error: "For 'integer' type, properties
maximum, minimum are not supported". Replace z.number().int() with
z.number() in the AI schema. The strict workoutSchema from core
validates integer constraints after the LLM response.

+++ packages/ai/src/adapters/ai-workout-schema.ts
@@ -15,7 +15,7 @@
 import { z } from "zod";
 
 const stepSchema = z.object({
-  stepIndex: z.number().int(),
+  stepIndex: z.number(),
   name: z.string().optional(),
   durationType: z.string(),
   duration: z.object({
@@ -43,7 +43,7 @@ const stepSchema = z.object({
 });
 
 const blockSchema = z.object({
-  repeatCount: z.number().int(),
+  repeatCount: z.number(),
   steps: z.array(stepSchema),
 });
